Fix: VS2010 compilation errors
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1530 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/avisplitter/AviFile.cpp
@@ -522,7 +522,7 @@ bool CAviFile::IsInterleaved(bool fKeepInfo)
 
 		curchunks[n]++;
 
-		if(cs2last.t >= 0 && abs(cs2min.n - cs2last.n) >= 1000)
+		if(cs2last.t >= 0 && abs((int)cs2min.n - (int)cs2last.n) >= 1000)
 			fInterleaved = false;
 
 		cs2last = cs2min;
